requirements.txt
click
pyperclip
setuptools
openai
python-dotenv
termcolor
prompt_toolkit
jsonpath_ng
